some refactor & make protect for some endpoints => next - create session for each qrcode

+++ backend/src/features/qrports/qrport.controller.ts
@@ -1,6 +1,8 @@
 import { IQRPort } from "@globalShared/types/entities/QRport.entity";
 import { QrPortService } from "./qrport.service";
 import { createEndpoint } from "@utils/basicApiFncs/createEndpoint";
+import { ConnectionService } from "@features/connections/connections.service";
+import { ApiError } from "@utils/basicApiFncs/ApiError";
 
 // create
 type CreateQrBody = { name: IQRPort["name"]; role: IQRPort["role"] };
@@ -13,7 +15,7 @@ type GetAllQrResponse = IQRPort[];
 type GetAllQrEndpoint = typeof createEndpoint<GetAllQrBody, GetAllQrResponse>;
 
 type DeleteQrPortBody = void;
-type DeleteQrPortResponse = { success: boolean };
+type DeleteQrPortResponse = void;
 type DeleteQrPortParams = { qrCode: IQRPort["qrCode"] };
 type DeleteQrPortEndpoint = typeof createEndpoint<
   DeleteQrPortBody,
@@ -22,25 +24,38 @@ type DeleteQrPortEndpoint = typeof createEndpoint<
 >;
 
 interface IQrPortController {
-  qrPortService: QrPortService;
   authEndpoint: typeof createEndpoint;
+  qrPortService: QrPortService;
+  connectionService: ConnectionService;
+
   getAll(): ReturnType<GetAllQrEndpoint>;
   create(): ReturnType<CreateQrEndpoint>;
   delete(): ReturnType<DeleteQrPortEndpoint>;
 }
 
 export class QrPortController implements IQrPortController {
-  qrPortService: QrPortService;
   authEndpoint: typeof createEndpoint;
+  qrPortService: QrPortService;
+  connectionService: ConnectionService;
 
-  constructor(qrPortService: QrPortService, authFn: typeof createEndpoint) {
-    this.qrPortService = qrPortService;
+  constructor(
+    authFn: typeof createEndpoint,
+    qrPortService: QrPortService,
+    connectionService: ConnectionService,
+  ) {
     this.authEndpoint = authFn;
+
+    this.qrPortService = qrPortService;
+    this.connectionService = connectionService;
   }
 
   getAll() {
     return this.authEndpoint<GetAllQrBody, GetAllQrResponse>(
       async (req, userUniqId) => {
+        if (!this.connectionService.hasRole(userUniqId, ["admin"])) {
+          throw new ApiError(403, "Forbidden");
+        }
+
         const all = this.qrPortService.getAll();
         return all;
       },
@@ -50,23 +65,35 @@ export class QrPortController implements IQrPortController {
   create() {
     return this.authEndpoint<CreateQrBody, CreateQrResponse>(
       async (req, userUniqId) => {
+        if (!this.connectionService.hasRole(userUniqId, ["admin"])) {
+          throw new ApiError(403, "Forbidden");
+        }
+
         const { name, role } = req.body;
+        if (!name || !role) {
+          throw new ApiError(400, "Missing name or role");
+        }
         const newQrPort = this.qrPortService.createQrPort(name, role);
         return newQrPort;
       },
     );
   }
 
-  // todo!!!
   delete() {
     return this.authEndpoint<
       DeleteQrPortBody,
       DeleteQrPortResponse,
       DeleteQrPortParams
     >(async (req, userUniqId) => {
+      if (!this.connectionService.hasRole(userUniqId, ["admin"])) {
+        throw new ApiError(403, "Forbidden");
+      }
+
       const { qrCode } = req.params;
       const result = this.qrPortService.delete(qrCode);
-      return { success: result };
+      if (!result) {
+        throw new ApiError(404, "QR-code not found");
+      }
     });
   }
 }
